+++ ymir/command/mir/commands/filter.py
@@ -5,7 +5,7 @@ from typing import Optional, Set
 from mir.commands import base
 from mir.protos import mir_command_pb2 as mirpb
 from mir.tools import annotations, checker, class_ids
-from mir.tools import mir_repo_utils, mir_storage_ops, revs_parser
+from mir.tools import mir_repo_utils, mir_storage, mir_storage_ops, revs_parser
 from mir.tools.code import MirCode
 from mir.tools.command_run_in_out import command_run_in_out
 from mir.tools.errors import MirRuntimeError
@@ -89,7 +89,7 @@ class CmdFilter(base.BaseCommand):
          _] = mir_storage_ops.MirStorageOps.load_multiple_storages(mir_root=mir_root,
                                                                    mir_branch=src_typ_rev_tid.rev,
                                                                    mir_task_id=src_typ_rev_tid.tid,
-                                                                   ms_list=mir_storage_ops.get_all_mir_storage(),
+                                                                   ms_list=mir_storage.get_all_mir_storage(),
                                                                    as_dict=False)
         task_id = dst_typ_rev_tid.tid
