+++ ymir/command/mir/tools/mir_repo_utils.py
@@ -4,9 +4,7 @@ from typing import List, Optional
 
 from mir import scm
 
-
-_MIR_FILES = ('metadatas.mir', 'annotations.mir', 'keywords.mir', 'tasks.mir', 'context.mir')
-_MIR_ASSOCIATED_FILES = ('.git', '.gitattributes', '.gitignore', '.mir', '.mir_lock')
+from mir.tools import mir_storage
 
 
 def find_extra_items(mir_root: str) -> List[str]:
@@ -14,7 +12,7 @@ def find_extra_items(mir_root: str) -> List[str]:
     find all extra items not in mir_settings.MIR_FILES_LIST
     """
     items = os.listdir(path=mir_root)
-    return list(set(items) - set(_MIR_FILES) - set(_MIR_ASSOCIATED_FILES))
+    return list(set(items) - set(mir_storage.get_all_mir_paths()) - set(mir_storage.MIR_ASSOCIATED_FILES))
 
 
 def mir_check_repo_git_dirty(mir_root: str = ".") -> bool:
